refactor(ecdh): refactor: mocha test files, test file
refactor: mocha test files, test file

re N
