+++ .mocharc.json
@@ -1,0 +1,6 @@
+{
+  "require": "ts-node/register",
+  "spec": "packages/ecdh/tests/**/*.test.ts",
+  "timeout": 20000,
+  "recursive": true
+}
